backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.6.0
python-decouple==3.8
dj-database-url==2.1.0
drf-spectacular==0.28.0
Pillow==11.0.0
gunicorn==23.0.0
whitenoise==6.8.2
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
